# server/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task API",
    "version": "1.0.0",
    "description": "In-memory Task Management API"
  },
  "paths": {
    "/": {
      "get": {
        "summary": "API Info Front Door",
        "responses": {
          "200": { "description": "API metadata" }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health Check Endpoint",
        "responses": {
          "200": { "description": "Server health status" }
        }
      }
    },
    "/tasks": {
      "get": {
        "summary": "Get all tasks",
        "responses": {
          "200": { "description": "List of all tasks" }
        }
      },
      "post": {
        "summary": "Create a new task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" }
                },
                "required": ["title"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Task created successfully" },
          "400": { "description": "Bad Request - Title required" }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Get task by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Task object" },
          "404": { "description": "Task not found" }
        }
      },
      "put": {
        "summary": "Update task by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "done": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Task updated successfully" },
          "404": { "description": "Task not found" }
        }
      },
      "delete": {
        "summary": "Delete task by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Task deleted successfully" },
          "404": { "description": "Task not found" }
        }
      }
    }
  }
}